+++ FanFollow/FanFollow/Source/Data/Protocol/LikeService.swift
@@ -1,0 +1,13 @@
+//
+//  LikeService.swift
+//  FanFollow
+//
+//  Created by parkhyo on 2023/07/04.
+//
+
+import Foundation
+import RxSwift
+
+protocol LikeService {
+    func fetchPostLike(id: String) -> Observable<[LikeDTO]>
+}
